requirements.txt
# Selenium WebDriver for browser automation
selenium==4.17.2

# Testing framework
pytest==8.0.0
pytest-html==4.1.1
pytest-xdist==3.5.0

# Test data generation
Faker==22.6.0

# Reporting and logging
allure-pytest==2.13.2
pytest-json-report==1.5.0

# Performance testing
locust==2.20.0

# Utilities
python-dotenv==1.0.0
openpyxl==3.1.2

# WebDriver Manager (auto-downloads drivers)
webdriver-manager==4.0.1
